# Product-Catalog-Clone/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Sistema Revendedora de Perfumes
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: active
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # Consultants
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, brand, category, price, imageUrl, active, createdAt, updatedAt]
      properties:
        id: { type: number }
        name: { type: string }
        brand: { type: string }
        category: { type: string }
        description: { type: ["string", "null"] }
        price: { type: number, description: "Price in cents" }
        originalPrice: { type: ["number", "null"], description: "Original price in cents before discount" }
        imageUrl: { type: string }
        active: { type: boolean }
        createdAt: { type: string }
        updatedAt: { type: string }
    ProductInput:
      type: object
      required: [name, brand, category, price, imageUrl]
      properties:
        name: { type: string, minLength: 1 }
        brand: { type: string, minLength: 1 }
        category: { type: string, minLength: 1 }
        description: { type: string }
        price: { type: number, minimum: 0 }
        originalPrice: { type: number, minimum: 0 }
        imageUrl: { type: string }
        active: { type: boolean, default: true }
    ProductUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        brand: { type: string, minLength: 1 }
        category: { type: string, minLength: 1 }
        description: { type: string }
        price: { type: number, minimum: 0 }
        originalPrice: { type: ["number", "null"] }
        imageUrl: { type: string }
        active: { type: boolean }
...