# Product-Catalog-Clone/lib/api-zod/src/generated/types/productUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Sistema Revendedora de Perfumes
 * OpenAPI spec version: 0.1.0
 */

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  brand?: string;
  /** @minLength 1 */
  category?: string;
  description?: string;
  /** @minimum 0 */
  price?: number;
  /** @nullable */
  originalPrice?: number | null;
  imageUrl?: string;
  active?: boolean;
}